update spdlog to 1.15.1

+++ src/transport/WebXClientCommandCollector.cpp
@@ -94,7 +94,7 @@ void WebXClientCommandCollector::mainLoop() {
                     // Deserialize instruction
                     auto instruction = this->_serializer->deserialize((const unsigned char *)message.data(), message.size());
                     if (instruction != NULL) {
-                        spdlog::debug("Received instruction type {}", instruction->type);
+                        spdlog::debug("Received instruction type {}", (int)instruction->type);
                         // Handle message
                         WebXController::instance()->onClientInstruction(instruction);
                     }
